+++ src/main/java/io/github/astrapi69/json/JsonFileToObjectExtensions.java
@@ -31,6 +31,8 @@ import java.util.Objects;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import io.github.astrapi69.json.factory.ObjectMapperFactory;
+
 /**
  * The class {@link JsonFileToObjectExtensions} converts json strings to java object and java
  * collections.
@@ -39,6 +41,24 @@ public final class JsonFileToObjectExtensions
 {
 	private JsonFileToObjectExtensions()
 	{
+	}
+
+	/**
+	 * Transforms the given json file into a java object.
+	 *
+	 * @param <T>
+	 *            the generic type
+	 * @param jsonFile
+	 *            the json file
+	 * @param clazz
+	 *            the clazz
+	 * @return the t
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred
+	 */
+	public static <T> T toObject(final File jsonFile, final Class<T> clazz) throws IOException
+	{
+		return toObject(jsonFile, clazz, ObjectMapperFactory.newObjectMapper());
 	}
 
 	/**